Refactor content script into modular structure
