+++ src/content/constants.js
@@ -1,0 +1,1 @@
+export { PDF_CONSTANTS, UI_CONSTANTS } from '../shared/constants.js';
