+++ src/content/ui.js
@@ -1,0 +1,1 @@
+export { PDFMonitorUI, showStandaloneError } from '../ui/PDFMonitorUI.js';
